+++ data/synthetic/sql/reporting/USP_IP_Sepsis_Details.sql
@@ -304,7 +304,7 @@ SELECT [PatientName] [PATIENTS]
 
 	, [ECMOYN] [ECMO Y/N]
 
-	, [IPSOSevereSepsisYN] [IPSO Severe Sepsis Y/N]
+	, [IPSOSevereSepsisYN] [SSS Severe Sepsis Y/N]
 
 	, [AlertNotActivatedReason] [CLINICAL_ALERTS Not Activated Reason]